IOS-4996 Updated key flow of settings
