--- Anytype/Sources/PresentationLayer/Settings/Account/SettingsAccountModuleOutput.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-@MainActor
-protocol SettingsAccountModuleOutput: AnyObject {
-    func onRecoveryPhraseSelected()
-    func onLogoutSelected()
-    func onDeleteAccountSelected()
-}
