+++ Anytype/Sources/PresentationLayer/Settings/Coordinator/SettingsCoordinatorView.swift
@@ -16,24 +16,21 @@ struct SettingsCoordinatorView: View {
             .sheet(isPresented: $model.showPushNotificationsSettings) {
                 PushNotificationsSettingsView()
             }
-            .sheet(isPresented: $model.showSettingsAccount) {
-                SettingsAccountView(output: model)
-                    .anytypeSheet(isPresented: $model.showLogoutAlert) {
-                        DashboardLogoutAlert {
-                            model.onBackupTap()
-                        } onLogout: {
-                            model.onLogoutConfirmTap()
-                        }
-                    }
-                    .anytypeSheet(isPresented: $model.showDeleteAccountAlert) {
-                        DashboardAccountDeletionAlert()
-                    }
-                    .sheet(isPresented: $model.showKeychainPhraseForLogout) {
-                        KeychainPhraseView(context: .logout)
-                    }
-                    .sheet(isPresented: $model.showKeychainPhraseForSettings) {
-                        KeychainPhraseView(context: .settings)
-                    }
+            .anytypeSheet(isPresented: $model.showLogoutAlert) {
+                DashboardLogoutAlert {
+                    model.onBackupTap()
+                } onLogout: {
+                    model.onLogoutConfirmTap()
+                }
+            }
+            .anytypeSheet(isPresented: $model.showDeleteAccountAlert) {
+                DashboardAccountDeletionAlert()
+            }
+            .sheet(isPresented: $model.showKeychainPhraseForLogout) {
+                KeychainPhraseView(context: .logout)
+            }
+            .sheet(isPresented: $model.showKeychainPhraseForSettings) {
+                KeychainPhraseView(context: .settings)
             }
             .sheet(isPresented: $model.showAbout) {
                 AboutView(output: model)
