+++ Anytype/Sources/PresentationLayer/Settings/Settings/Sections/SettingsSectionItemView.swift
@@ -9,21 +9,24 @@ extension SettingsSectionItemView {
 
 struct SettingsSectionItemView: View {
     let name: String
+    let textColor: Color
     let iconImage: Icon?
     let decoration: Decoration?
     let showDivider: Bool
     let onTap: () -> Void
         
-    init(name: String, imageAsset: ImageAsset, decoration: Decoration? = .arrow(), showDivider: Bool = true, onTap: @escaping () -> Void) {
+    init(name: String, textColor: Color = .Text.primary, imageAsset: ImageAsset, decoration: Decoration? = .arrow(), showDivider: Bool = true, onTap: @escaping () -> Void) {
         self.name = name
+        self.textColor = textColor
         self.iconImage = .asset(imageAsset)
         self.decoration = decoration
         self.showDivider = showDivider
         self.onTap = onTap
     }
     
-    init(name: String, iconImage: Icon? = nil, decoration: Decoration? = .arrow(), showDivider: Bool = true, onTap: @escaping () -> Void) {
+    init(name: String, textColor: Color = .Text.primary, iconImage: Icon? = nil, decoration: Decoration? = .arrow(), showDivider: Bool = true, onTap: @escaping () -> Void) {
         self.name = name
+        self.textColor = textColor
         self.iconImage = iconImage
         self.decoration = decoration
         self.showDivider = showDivider
@@ -43,7 +46,7 @@ struct SettingsSectionItemView: View {
                 }
                 HStack(alignment: .center, spacing: 0) {
                     AnytypeText(name, style: .uxBodyRegular)
-                        .dynamicForegroundStyle(color: .Text.primary, disabledColor: .Text.tertiary)
+                        .dynamicForegroundStyle(color: textColor, disabledColor: .Text.tertiary)
                     Spacer()
                     decorationView
                 }
